Moved templated node finder to header file so that it could be instantiated properly in tests.
